some refactoring, however this breaks the case "echo a | wazero run ex.wasm 0"
where 0 is the timeout. Even with timeout 0, stdin should be
considered ready because we are piping in some readily-available data
(it is not a tty)

however, the timer expires before we get a chance to write
back the value to the buffer.

The previous commit was working by executing some parts
of the code deterministically (not in a go routine context),
but in real life, the execution was flaky! Sometimes it returned
as ready, sometimes as no-data!

Signed-off-by: Edoardo Vacchi <evacchi@users.noreply.github.com>