packaging: switch to native WiX toolchain and Windows runner
Cross-platform msitools do not support the `Environment` directive,
which should be the proper way to address the issue, so we adapt
the Makefile to assume a Windows environment when building, addressing #1374.

`make dist` now runs on all platforms, but on Windows, assuming
`candle.exe`, `light.exe` and GNU `tar` are installed, it builds
all tarballs, zip and msi packages.

On other platforms it will only build the tarballs and print a warning:

    Makefile:340: WARNING: MSI generation only supported on Windows platforms. Skipping.

Signed-off-by: Edoardo Vacchi <evacchi@users.noreply.github.com>