wasi: nonblocking I/O for sockets and pipes on Windows
Further work to improve the support to nonblocking I/O on Windows.
This drops the special-casing for stdin, and instead checks if
a handle is a named pipe (which includes sockets and other
kinds of pipes); moreover it further improve _select by
introducing a proper wrapper for winsock's select, which
is BSD socket's select, and a compatible implementation of FdSet.

Signed-off-by: Edoardo Vacchi <evacchi@users.noreply.github.com>